Move functions from index to individual files and isolate tests.
