+++ package.json
@@ -42,7 +42,6 @@
     "lint-staged": "^7.0.5",
     "prettier": "1.12.1",
     "ts-jest": "^22.4.4",
-    "ts-loader": "^4.2.0",
     "tslint": "^5.9.1",
     "tslint-immutable": "^4.5.4",
     "typescript": "^2.8.3",
