+++ src/globals/on-install.ts
@@ -1,0 +1,11 @@
+import { addMenu } from "../functions";
+
+/**
+ * Runs when the add-on is installed to a spreadsheet.
+ *
+ * @export
+ * @returns {void}
+ */
+export function onInstall(): void {
+  return addMenu();
+}
