+++ src/globals/on-open.ts
@@ -1,0 +1,11 @@
+import { addMenu } from "../functions";
+
+/**
+ * Runs when the spreadsheet is opened.
+ *
+ * @export
+ * @returns {void}
+ */
+export function onOpen(): void {
+  return addMenu();
+}
