+++ webpack.config.js
@@ -1,15 +1,51 @@
 const GasPlugin = require("gas-webpack-plugin");
 const UglifyJSPlugin = require("uglifyjs-webpack-plugin");
 
-module.exports = {
-  context: __dirname,
+const path = require("path");
+const basePath = __dirname;
+
+const presets = [
+  ["@babel/preset-typescript"],
+  [
+    "@babel/preset-env",
+    {
+      targets: {
+        browsers: ["ie >= 9"]
+      },
+      useBuiltIns: "usage",
+      modules: "commonjs",
+      debug: false
+    }
+  ]
+];
+
+const uglify = new UglifyJSPlugin({
+  uglifyOptions: {
+    compress: {
+      unused: false // TypeError: redeclaration of const undefined if true
+    },
+    mangle: false,
+    toplevel: false,
+    keep_classnames: true,
+    keep_fnames: true,
+    ie8: true,
+    ecma: 5,
+    output: {
+      comments: true,
+      beautify: true
+    }
+  }
+});
+
+const config = {
+  context: basePath,
   mode: "production",
   entry: {
-    sheets: "./src/index.ts"
+    sheets: path.join(basePath, "src", "index.ts")
   },
   output: {
-    path: __dirname,
-    filename: "./dist/[name].js"
+    path: path.join(basePath, "dist"),
+    filename: "[name].js"
   },
   plugins: [new GasPlugin()],
   resolve: {
@@ -19,52 +55,27 @@ module.exports = {
     rules: [
       {
         test: /\.ts(x?)$/,
+        exclude: /node_modules/,
         loaders: [
           {
-            loader: "babel-loader", // babel-loader runs last, transpiling to GAS compatible JS.
-            options: {
-              presets: [
-                ["@babel/preset-typescript"],
-                [
-                  "@babel/preset-env",
-                  {
-                    targets: {
-                      browsers: ["ie >= 9"]
-                    },
-                    useBuiltIns: "usage",
-                    debug: true
-                  }
-                ]
-              ]
-            }
+            loader: "babel-loader",
+            options: { presets }
           }
         ]
       },
       {
         test: /\.html$/,
+        exclude: /node_modules/,
         loader: "html-loader"
       }
     ]
   },
   optimization: {
-    // Production mode uses UglifyJSPlugin by default but we need to modify it to work with GAS.
-    minimizer: [
-      // Modifying this configuration will very likely break the script in GAS. Make sure to test!
-      new UglifyJSPlugin({
-        uglifyOptions: {
-          compress: false,
-          mangle: false,
-          toplevel: false,
-          keep_classnames: true,
-          keep_fnames: true,
-          ie8: true,
-          ecma: 5,
-          output: {
-            comments: false,
-            beautify: true
-          }
-        }
-      })
-    ]
+    minimizer: [uglify],
+    runtimeChunk: false,
+    noEmitOnErrors: true,
+    concatenateModules: true
   }
 };
+
+module.exports = config;
